Two scenarios implemented
1. In the first scenario, we’d work with a predefined-one-role schema.
This way we’d be able to modify this role in any fashion so our rules
are applied to this existing role.

2. In this scenario, we’d work with a custom RoleMapper in order to
tell mondrian, which roles to apply to the user. Then, we could use the
DSP to inject our existing roles so the rules would be applied to more
than one role.

In this second scenario, i only succeeded by adding to the xml schema a
pre-existing role named “dummy” with a schema-grant in it.