Fixed str.replacement and catalog name parameter.
The string to be replaced was never being effectively being passed to
the schema, as the replaced value wasn't being assigned to the output
variable.

Also, the datasource name was being passed to the endpoint, instead of
the Schema [catalog] name. So instead of getting the datasource, I
started pointing this var to the prop Catalog and also refactored the
var name everywhere into the code.

Also added a property called cubeguard.debug that when true will print
vars to the console.